reorg file structure and add version matching based on semver ranges
