+++ index.js
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 'use strict';
+var fs = require('fs');
 var path = require('path');
 var gutil = require('gulp-util');
 var chalk = require('chalk');
@@ -10,16 +11,20 @@ var Liftoff = require('liftoff');
 var tildify = require('tildify');
 var interpret = require('interpret');
 var v8flags = require('v8flags');
-var exit = require('./lib/exit');
-var cliOptions = require('./lib/cliOptions');
-var completion = require('./lib/completion');
-var verifyDeps = require('./lib/verifyDependencies');
+var findRange = require('semver-greatest-satisfied-range');
+var exit = require('./lib/shared/exit');
+var cliOptions = require('./lib/shared/cliOptions');
+var completion = require('./lib/shared/completion');
+var verifyDeps = require('./lib/shared/verifyDependencies');
 var cliVersion = require('./package.json').version;
-var getBlacklist = require('./lib/getBlacklist');
+var getBlacklist = require('./lib/shared/getBlacklist');
 
 // logging functions
-var logVerify = require('./lib/log/verify');
-var logBlacklistError = require('./lib/log/blacklistError');
+var logVerify = require('./lib/shared/log/verify');
+var logBlacklistError = require('./lib/shared/log/blacklistError');
+
+// get supported ranges
+var ranges = fs.readdirSync(__dirname + '/lib/versioned/');
 
 // set env var for ORIGINAL cwd
 // before anything touches it
@@ -36,8 +41,6 @@ var opts = nomnom
   .script('gulp')
   .options(cliOptions)
   .parse();
-var tasks = opts._;
-var toRun = tasks.length ? tasks : ['default'];
 
 // this is a hold-over until we have a better logging system
 // with log levels
@@ -115,6 +118,7 @@ function handleArguments(env) {
   }
 
   // check for semver difference between cli and local installation
+  // TODO: remove when we support all
   if (semver.gt(cliVersion, env.modulePackage.version)) {
     gutil.log(chalk.red('Warning: gulp version mismatch:'));
     gutil.log(chalk.red('Global gulp is', cliVersion));
@@ -131,75 +135,9 @@ function handleArguments(env) {
     );
   }
 
-  if (semver.satisfies(env.modulePackage.version, '^3')) {
-    return gulp3(env);
-  }
+  // find the correct CLI version to run
+  var range = findRange(env.modulePackage.version, ranges);
 
-  if (semver.satisfies(env.modulePackage.version, '^4')) {
-    return gulp4(env);
-  }
-}
-
-function gulp3(env){
-  var logTasks = require('./lib/3/logTasks');
-  var logEvents = require('./lib/3/logEvents');
-  var logTasksSimple = require('./lib/3/logTasksSimple');
-
-  // this is what actually loads up the gulpfile
-  require(env.configPath);
-  gutil.log('Using gulpfile', chalk.magenta(tildify(env.configPath)));
-
-  var gulpInst = require(env.modulePath);
-  logEvents(gulpInst);
-
-  process.nextTick(function () {
-    if (opts.tasksSimple) {
-      return logTasksSimple(env, gulpInst);
-    }
-    if (opts.tasks) {
-      return logTasks(env, gulpInst);
-    }
-    gulpInst.start.apply(gulpInst, toRun);
-  });
-}
-
-function gulp4(env){
-  var logTasks = require('./lib/log/tasks');
-  var logEvents = require('./lib/log/events');
-  var logTasksSimple = require('./lib/log/tasksSimple');
-
-  var gulpInst = require(env.modulePath);
-  logEvents(gulpInst);
-
-  // this is what actually loads up the gulpfile
-  require(env.configPath);
-
-  process.nextTick(function () {
-    if (opts.tasksSimple) {
-      return logTasksSimple(gulpInst.tree());
-    }
-    if (opts.tasks) {
-      var tree = {
-        label: 'Tasks for ' + chalk.magenta(tildify(env.configPath)),
-        nodes: gulpInst.tree({ deep: true })
-      };
-      return logTasks(tree);
-    }
-    if (opts.tasksJson) {
-      return console.log(
-        JSON.stringify(gulpInst.tree({ deep: true }), null, 2)
-      );
-    }
-    try {
-      gutil.log('Using gulpfile', chalk.magenta(tildify(env.configPath)));
-      // TODO: do we care about the error/result from calling this?
-      gulpInst.parallel(toRun)();
-    } catch (err) {
-      gutil.log(chalk.red(err.message));
-      gutil.log(
-        'Please check the documentation for proper gulpfile formatting'
-      );
-      exit(1);
-    }
-  });
+  // load and execute the CLI version
+  require(path.join(__dirname, '/lib/versioned/', range, '/'))(opts, env);
 }
