+++ lib/shared/cliOptions.js
