+++ lib/shared/completion.js
