+++ lib/shared/exit.js
