+++ lib/shared/getBlacklist.js
