+++ lib/shared/log/blacklistError.js
@@ -3,12 +3,12 @@
 var chalk = require('chalk');
 var gutil = require('gulp-util');
 
-var formatError = require('../formatError');
+var exit = require('../exit');
 
 function logBlacklistError(err) {
   gutil.log(chalk.red('Error: failed to retrieve plugins black-list'));
-  gutil.log(formatError(err));
-  process.exit(1);
+  gutil.log(err.message); // avoid duplicating for each version
+  exit(1);
 }
 
 module.exports = logBlacklistError;
