+++ lib/shared/log/verify.js
@@ -3,6 +3,8 @@
 var chalk = require('chalk');
 var gutil = require('gulp-util');
 
+var exit = require('../exit');
+
 function logVerify(blacklisted) {
   var pluginNames = Object.keys(blacklisted);
 
@@ -10,7 +12,7 @@ function logVerify(blacklisted) {
     gutil.log(
       chalk.green('There are no blacklisted plugins in this project')
     );
-    process.exit(0);
+    exit(0);
   }
 
   gutil.log(chalk.red('Blacklisted plugins found in this project:'));
@@ -20,7 +22,7 @@ function logVerify(blacklisted) {
     gutil.log(chalk.bgRed(pluginName) + ': ' + reason);
   });
 
-  process.exit(1);
+  exit(1);
 }
 
 module.exports = logVerify;
