+++ lib/shared/verifyDependencies.js
