+++ lib/versioned/^3.7.0/log/events.js
@@ -4,8 +4,8 @@ var chalk = require('chalk');
 var gutil = require('gulp-util');
 var prettyTime = require('pretty-hrtime');
 
-var exit = require('../exit');
-var formatError = require('./formatError');
+var exit = require('../../../shared/exit');
+var formatError = require('../formatError');
 
 // wire up logging events
 function logEvents(gulpInst) {
