+++ lib/versioned/^3.7.0/log/tasks.js
@@ -5,7 +5,7 @@ var chalk = require('chalk');
 var gutil = require('gulp-util');
 var tildify = require('tildify');
 
-var taskTree = require('./taskTree');
+var taskTree = require('../taskTree');
 
 function logTasks(env, localGulp) {
   var tree = taskTree(localGulp.tasks);
