+++ lib/versioned/^4.0.0/formatError.js
