+++ package.json
@@ -34,6 +34,7 @@
     "nomnom": "^1.8.0",
     "pretty-hrtime": "^0.2.0",
     "semver": "^4.1.0",
+    "semver-greatest-satisfied-range": "^1.0.0",
     "tildify": "^1.0.0",
     "v8flags": "^1.0.1",
     "wreck": "^5.0.1"
